pyproject.toml
[project]
name = "deepresearchtool"
version = "0.1.0"
description = "Advanced research tool optimized for M1 MacBook with deep web capabilities"
requires-python = ">=3.11"
dependencies = [
    # Core framework dependencies
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",

    # HTTP and networking
    "httpx==0.28.1",
    "aiohttp==3.12.15",
    "aiohttp-socks==0.10.1",
    "requests==2.32.5",
    "pysocks==1.7.1",

    # Retry and error handling (FÁZE 1)
    "tenacity==9.0.0",

    # Web scraping and automation
    "playwright==1.55.0",
    "playwright-stealth==2.0.0",
    "beautifulsoup4==4.13.5",

    # Anonymity and deep web access
    "stem==1.8.2",  # Tor control
    "i2plib==0.0.14",  # I2P support

    # Vector databases and search
    "chromadb==1.0.20",  # Primary vector DB for M1 optimization
    "qdrant-client==1.15.1",  # Alternative vector DB
    "lancedb==0.24.3",  # Alternative vector DB
    "rank-bm25==0.2.2",  # BM25 search

    # Graph database and network analysis
    "neo4j==5.28.2",
    "networkx==3.4.2",  # Graph analysis and network algorithms

    # LangChain ecosystem
    "langchain-core==0.3.29",
    "langchain-openai==0.2.14",
    "openai==1.57.0",

    # System monitoring and utilities
    "psutil==6.1.1",  # System and process utilities
    "firecrawl-py==1.7.2",  # Web scraping service

    # Data processing and storage
    "polars==1.32.3",  # Fast DataFrame library optimized for M1
    "pyarrow==21.0.0",
    "duckdb==1.3.2",  # Fast analytical database

    # Task queue and workers
    "dramatiq==1.18.0",
    "redis==6.4.0",

    # Configuration and utilities
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "python-dateutil==2.9.0.post0",

    # Streaming and async processing
    "ijson==3.3.0",  # Streaming JSON parser for large files

    # UI and visualization
    "streamlit==1.49.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "pytest-xdist==3.6.0",  # Parallel test execution

    # Code quality
    "ruff==0.8.4",  # Linting and formatting
    "mypy==1.13.0",  # Type checking
    "pre-commit==4.0.1",

    # Security - removed safety for now due to pydantic conflicts
    "bandit[toml]==1.8.0",  # Security linting
    # "safety==3.2.7",  # Temporarily disabled due to pydantic version conflicts

    # Documentation
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.47",
]

m1-optimized = [
    # M1-specific optimizations
    "accelerate==1.2.1",  # Hugging Face acceleration for M1
    "torch==2.5.1",  # PyTorch with M1 Metal support
    "torchvision==0.20.1",
    "torchaudio==2.5.1",
]

production = [
    # Production monitoring and observability
    "prometheus-client==0.21.1",
    "structlog==24.4.0",
    "sentry-sdk[fastapi]==2.19.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--asyncio-mode=auto"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "external: marks tests that require external services",
    "ai_dependent: marks tests that require AI/LLM services",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "stem.*",
    "i2plib.*",
    "playwright_stealth.*",
    "dramatiq.*",
    "qdrant_client.*",
    "chromadb.*",
    "lancedb.*",
    "polars.*",
]
ignore_missing_imports = true
